# backend/ent/schema/unit.go
package schema

import (
	"github.com/facebookincubator/ent"
	"github.com/facebookincubator/ent/schema/field" 
	"github.com/facebookincubator/ent/schema/edge"
)

// Unit holds the schema definition for the User entity.
type Unit struct {
	ent.Schema
}

// Fields of the User.
func (Unit) Fields() []ent.Field {
	return []ent.Field {
		field.String("UnitType").NotEmpty(), 
		
	}
}

// Edges of the User.
func (Unit) Edges() []ent.Edge {
	return []ent.Edge {
		edge.To("drugs",Drug.Type).StorageKey(edge.Column("unit_id")),
	}
}